+++ client/src/components/Pages/AddReview/ReviewSection/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import ReviewWrapper from "../Wrappers/ReviewWrapper";
+
+const ReviewSection = () => {
+  return (
+      <ReviewWrapper>
+        <h1>ReviewSection</h1>
+        <h1>Kurwa</h1>
+      </ReviewWrapper>
+  );
+};
+
+export default ReviewSection;
