+++ client/src/components/Pages/AddReview/Wrappers/ReviewWrapper.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const ReviewWrapper = styled.section`
+  padding: 0 0 0 12.5rem;
+  display: flex;
+  flex-direction: row;
+`;
+
+export default ReviewWrapper;
