+++ client/src/components/Pages/AddReview/index.js
@@ -2,8 +2,8 @@ import React from "react";
 
 import Avatar from "./Avatar";
 import CallForReview from "./CallForReview";
-import ComponentWrapper from "./ComponentWrapper";
-import ReviewSection from "./ReviewForm";
+import ComponentWrapper from "./Wrappers/ComponentWrapper";
+import ReviewSection from "./ReviewSection";
 
 const AddReview = () => {
   return (
